+++ iota-client/src/node_manager.rs
@@ -579,6 +579,7 @@ pub(crate) struct HttpClient {
 }
 
 #[cfg(all(feature = "sync", not(feature = "async")))]
+#[derive(Clone)]
 pub(crate) struct HttpClient;
 
 #[cfg(any(feature = "async", feature = "wasm"))]