+++ iota-client/tests/addresses.rs
@@ -1,7 +1,7 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use iota_client::{api::GetAddressesBuilder, Seed};
+use iota_client::{api::GetAddressesBuilder, Client, Seed};
 
 #[tokio::test]
 async fn addresses() {
@@ -24,4 +24,19 @@ async fn addresses() {
         "atoi1qprxpfvaz2peggq6f8k9cj8zfsxuw69e4nszjyv5kuf8yt70t2847shpjak".to_string()
     );
     assert!(addresses[1].1);
+}
+#[tokio::test]
+async fn public_key_to_address() {
+    let iota = Client::builder().with_offline_mode().finish().await.unwrap();
+
+    let hex_public_key = "2baaf3bca8ace9f862e60184bd3e79df25ff230f7eaaa4c7f03daa9833ba854a";
+
+    let public_key_address = iota
+        .hex_public_key_to_bech32_address(hex_public_key, Some("atoi"))
+        .await
+        .unwrap();
+    assert_eq!(
+        public_key_address,
+        "atoi1qzt0nhsf38nh6rs4p6zs5knqp6psgha9wsv74uajqgjmwc75ugupx3y7x0r".to_string()
+    );
 }